Image block: split huge component
